Moving rx-java2 to test dependency [from compile-time dependency].
