+++ client-runtime/src/main/java/com/microsoft/rest/v3/RestProxy.java
@@ -31,7 +31,7 @@ import com.microsoft.rest.v3.serializer.JacksonAdapter;
 import com.microsoft.rest.v3.util.FluxUtil;
 import com.microsoft.rest.v3.util.TypeUtil;
 import io.netty.buffer.ByteBuf;
-import io.reactivex.exceptions.Exceptions;
+import reactor.core.Exceptions;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
